mainviewmodel partially done

+++ app/src/main/java/com/rsstudio/weather/data/network/model/Weather.kt
@@ -7,8 +7,8 @@ data class Weather(
     val generationtime_ms: Double,
     val hourly: Hourly,
     val hourly_units: HourlyUnits,
-    val latitude: Int,
-    val longitude: Int,
+    val latitude: Double,
+    val longitude: Double,
     val timezone: String,
     val timezone_abbreviation: String,
     val utc_offset_seconds: Int
